requirements.txt
python-dotenv==0.21.0
psycopg2-binary==2.9.5
requests==2.28.2
beautifulsoup4==4.11.1
